Change unittest with log. add somme dll

+++ PetProject.Core/Entities/BaseEntity.cs
@@ -10,9 +10,9 @@ namespace PetProject.Core.Entities
     {
         public long Id { get; set; }
         public DateTime CreatedTime { get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public DateTime? UpdatedTime { get; set; }
-        public int? UpdatedBy { get; set; }
+        public long? UpdatedBy { get; set; }
         public byte[]? RowVersion { get; set; }
     }
 }
